Start removal of asdf
Even though asdf looks to be a very useful tool, it unfortunately has a
fatal flaw related performance. Each invocation of any asdf-managed
tool incurs at 300-400ms delay.

I know that probably doesn't sound like much, but some scripts that make
dozens of invocations (to perl, for example) easily take upwards of 20-30
seconds to load when using asdf (and about 10ms without it).

There's a direnv plugin that tries to solve the problem (and actually
performs quite well), but requires littering `.envrc` files throughout
the file system (in addition to the `.tool-versions` and other files
that asdf already requires). And seeing as I was using asdf in an effort
to consolidate tools and reduce clutter, this is a no-go.

I'm very sad to have to say goodbye, but hopefully someday they will fix
this extremely detrimental performance problem (and I'll be first back
in line)

See: https://github.com/asdf-vm/asdf/issues/290