Pin TrueZen to last commit before plugin rewrite
This is potentially a temporary pin until I have a chance to go through
the new settings and update, but maybe it's permanent. The plugin rewrite
appears to have taken out a lot of functionality that I depend on, so
it'll likely stay pinned until it stops working, and then hopefully
another plugin will exist to switch to by then.