override get_root function for lazyvim
Lazyvim has this whole functionality that overrides the root dir
from the terminal in favor of a bunch of stuff that I dislike. This is
baked in to every corner of lazyvim, so rather than overriding each
keymap, here we override the function to just return the cwd from the
terminal, and keep everything else.

Idea from:
https://github.com/LazyVim/LazyVim/issues/704#issuecomment-1533378129